created a go modules file and adjusted references
