+++ go.mod
@@ -1,0 +1,15 @@
+module physarum
+
+go 1.15
+
+require (
+	github.com/go-gl/gl v0.0.0-20210315015930-ae072cafe09d
+	github.com/go-gl/glfw v0.0.0-20210311203641-62640a716d48
+	github.com/gonum/blas v0.0.0-20181208220705-f22b278b28ac // indirect
+	github.com/gonum/floats v0.0.0-20181209220543-c233463c7e82 // indirect
+	github.com/gonum/integrate v0.0.0-20181209220457-a422b5c0fdf2 // indirect
+	github.com/gonum/internal v0.0.0-20181124074243-f884aa714029 // indirect
+	github.com/gonum/lapack v0.0.0-20181123203213-e4cdc5a0bff9 // indirect
+	github.com/gonum/matrix v0.0.0-20181209220409-c518dec07be9 // indirect
+	github.com/gonum/stat v0.0.0-20181125101827-41a0da705a5b
+)
