+++ physarum/blur.go
