+++ physarum/config.go
