+++ physarum/grid.go
