+++ physarum/particle.go
