+++ physarum/trig.go
