+++ physarum/trig_test.go
