+++ physarum/util.go
